requirements.txt
crewai
crewai-tools
fastapi
uvicorn[standard]
streamlit
python-dotenv
requests
pydantic